Changes: 1. Layouts are improved and each is moved to a separate class 2. Graph diff new api is defined. 3. Logging is added to VertexFilter Factory (system.out is replaced). 4. New and OpenProject stuff are changed in a way that now discovery, bgpPeeringMap and freeGraph projects could be created and managed. 5. Topology Viewer matchers are done to work with long instead of int (we got a use case in bgpPeeringMaps that required that). 6. BigBuffered image has been added since we got an issue with too big images in bgp peering map generation 7. viewer configs were improved so that it points in a good way paths to fulfillment & parameter factories were changed. 8. Parameter and fulfillment factory configurations were added to discovery projects. 9. CliDiscovery launcher was enhanced!

+++ iTopologyManager/topologyViewerConfigAPI/src/main/java/net/itransformers/topologyviewer/config/models/datamatcher/impl/BetweenOrEqualInterval.java
@@ -51,9 +51,9 @@ public class BetweenOrEqualInterval implements DataMatcher{
                 logger.error("Unable to parse input limits: "+data2);
                 return false;
             }
-            Integer i11 = Integer.parseInt(limits[0]);
-            Integer i12 = Integer.parseInt(limits[1]);
-            Integer i2 = Integer.parseInt(data1);
+            Long i11 = Long.parseLong(limits[0]);
+            Long i12 = Long.parseLong(limits[1]);
+            Long i2 = Long.parseLong(data1);
             logger.debug("beoq: "+i11+":"+i2+":"+i12);
 
             return (i2 >= i11 && i2 <= i12);
